+++ mcs/class/System/System.Net.NetworkInformation/LinuxNetworkInterfaceMarshal.cs
@@ -25,7 +25,6 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-using System;
 using System.Runtime.InteropServices;
 
 namespace System.Net.NetworkInformation {
