+++ mcs/class/System/System.Net.NetworkInformation/MacOsNetworkInterfaceMarshal.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Runtime.InteropServices;
 
 namespace System.Net.NetworkInformation {
