+++ mcs/class/System/System.Net.NetworkInformation/UnixIPGlobalProperties.cs
@@ -27,12 +27,10 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Globalization;
 using System.IO;
-using System.Net.Sockets;
 using System.Runtime.InteropServices;
 using System.Text;
 
@@ -358,342 +356,38 @@ namespace System.Net.NetworkInformation {
 		}
 	}
 
-#if WIN_PLATFORM
-	class Win32IPGlobalProperties : IPGlobalProperties
-	{
-		public const int AF_INET = 2;
-		public const int AF_INET6 = 23;
-
-		// FIXME: it might be getting wrong table. I'm getting
-		// different results from .NET 2.0.
-		unsafe void FillTcpTable (out List<Win32_MIB_TCPROW> tab4, out List<Win32_MIB_TCP6ROW> tab6)
+	internal static class UnixIPGlobalPropertiesFactoryPal {
+		public static IPGlobalProperties Create ()
 		{
-			tab4 = new List<Win32_MIB_TCPROW> ();
-			int size4 = 0;
-			GetTcpTable (null, ref size4, true); // get size
-			byte [] bytes4 = new byte [size4];
-			GetTcpTable (bytes4, ref size4, true); // get list
-
-			int structSize4 = Marshal.SizeOf (typeof (Win32_MIB_TCPROW));
-
-			fixed (byte* ptr = bytes4) {
-				int count = Marshal.ReadInt32 ((IntPtr) ptr);
-				for (int i = 0; i < count; i++) {
-					Win32_MIB_TCPROW row = new Win32_MIB_TCPROW ();
-					Marshal.PtrToStructure ((IntPtr) (ptr + i * structSize4 + 4), row);
-					tab4.Add (row);
+#if MONODROID
+			return new AndroidIPGlobalProperties ();
+#elif MONOTOUCH || XAMMAC
+			return new UnixIPGlobalProperties ();
+#elif MONO
+			switch (Environment.OSVersion.Platform) {
+			case PlatformID.Unix:
+				MibIPGlobalProperties impl = null;
+				if (Directory.Exists (MibIPGlobalProperties.ProcDir)) {
+					impl = new MibIPGlobalProperties (MibIPGlobalProperties.ProcDir);
+					if (File.Exists (impl.StatisticsFile))
+						return impl;
 				}
-			}
-
-			tab6 = new List<Win32_MIB_TCP6ROW> ();
-			if (Environment.OSVersion.Version.Major >= 6) { // Vista
-				int size6 = 0;
-				GetTcp6Table (null, ref size6, true); // get size
-				byte [] bytes6 = new byte [size6];
-				GetTcp6Table (bytes6, ref size6, true); // get list
-
-				int structSize6 = Marshal.SizeOf (typeof (Win32_MIB_TCP6ROW));
-
-				fixed (byte* ptr = bytes6) {
-					int count = Marshal.ReadInt32 ((IntPtr) ptr);
-					for (int i = 0; i < count; i++) {
-						Win32_MIB_TCP6ROW row = new Win32_MIB_TCP6ROW ();
-						Marshal.PtrToStructure ((IntPtr) (ptr + i * structSize6 + 4), row);
-						tab6.Add (row);
-					}
+				if (Directory.Exists (MibIPGlobalProperties.CompatProcDir)) {
+					impl = new MibIPGlobalProperties (MibIPGlobalProperties.CompatProcDir);
+					if (File.Exists (impl.StatisticsFile))
+						return impl;
 				}
-			}
+				return new UnixIPGlobalProperties ();
+			default:
+#if !WIN_PLATFORM
+				return new UnixIPGlobalProperties ();
+#endif
+				return null;
 		}
-
-		bool IsListenerState (TcpState state)
-		{
-			switch (state) {
-			case TcpState.SynSent:
-			case TcpState.Listen:
-			case TcpState.FinWait1:
-			case TcpState.FinWait2:
-			case TcpState.CloseWait:
-				return true;
-			}
-			return false;
-		}
-
-		public override TcpConnectionInformation [] GetActiveTcpConnections ()
-		{
-			List<Win32_MIB_TCPROW> tab4 = null;
-			List<Win32_MIB_TCP6ROW> tab6 = null;
-			FillTcpTable (out tab4, out tab6);
-			int size4 = tab4.Count;
-
-			TcpConnectionInformation [] ret = new TcpConnectionInformation [size4 + tab6.Count];
-			for (int i = 0; i < size4; i++)
-				ret [i] = tab4 [i].TcpInfo;
-			for (int i = 0; i < tab6.Count; i++)
-				ret [size4 + i] = tab6 [i].TcpInfo;
-			return ret;
-		}
-
-		public override IPEndPoint [] GetActiveTcpListeners ()
-		{
-			List<Win32_MIB_TCPROW> tab4 = null;
-			List<Win32_MIB_TCP6ROW> tab6 = null;
-			FillTcpTable (out tab4, out tab6);
-
-			List<IPEndPoint> ret = new List<IPEndPoint> ();
-			for (int i = 0, count = tab4.Count; i < count; i++)
-				if (IsListenerState (tab4 [i].State))
-					ret.Add (tab4 [i].LocalEndPoint);
-			for (int i = 0, count = tab6.Count; i < count; i++)
-				if (IsListenerState (tab6 [i].State))
-					ret.Add (tab6 [i].LocalEndPoint);
-			return ret.ToArray ();
-		}
-
-		public unsafe override IPEndPoint [] GetActiveUdpListeners ()
-		{
-			List<IPEndPoint> list = new List<IPEndPoint> ();
-
-			byte [] bytes4 = null;
-			int size4 = 0;
-			GetUdpTable (null, ref size4, true); // get size
-			bytes4 = new byte [size4];
-			GetUdpTable (bytes4, ref size4, true); // get list
-
-			int structSize4 = Marshal.SizeOf (typeof (Win32_MIB_UDPROW));
-
-			fixed (byte* ptr = bytes4) {
-				int count = Marshal.ReadInt32 ((IntPtr) ptr);
-				for (int i = 0; i < count; i++) {
-					Win32_MIB_UDPROW row = new Win32_MIB_UDPROW ();
-					Marshal.PtrToStructure ((IntPtr) (ptr + i * structSize4 + 4), row);
-					list.Add (row.LocalEndPoint);
-				}
-			}
-
-			if (Environment.OSVersion.Version.Major >= 6) { // Vista
-				byte [] bytes6 = null;
-				int size6 = 0;
-				GetUdp6Table (null, ref size6, true); // get size
-				bytes6 = new byte [size6];
-				GetUdp6Table (bytes6, ref size6, true); // get list
-
-				int structSize6 = Marshal.SizeOf (typeof (Win32_MIB_UDP6ROW));
-
-				fixed (byte* ptr = bytes6) {
-					int count = Marshal.ReadInt32 ((IntPtr) ptr);
-					for (int i = 0; i < count; i++) {
-						Win32_MIB_UDP6ROW row = new Win32_MIB_UDP6ROW ();
-						Marshal.PtrToStructure ((IntPtr) (ptr + i * structSize6 + 4), row);
-						list.Add (row.LocalEndPoint);
-					}
-				}
-			}
-
-			return list.ToArray ();
-		}
-
-		public override IcmpV4Statistics GetIcmpV4Statistics ()
-		{
-			if (!Socket.OSSupportsIPv4)
-				throw new NetworkInformationException ();
-			Win32_MIBICMPINFO stats;
-			GetIcmpStatistics (out stats, AF_INET);
-			return new Win32IcmpV4Statistics (stats);
-		}
-
-		public override IcmpV6Statistics GetIcmpV6Statistics ()
-		{
-			if (!Socket.OSSupportsIPv6)
-				throw new NetworkInformationException ();
-			Win32_MIB_ICMP_EX stats;
-			GetIcmpStatisticsEx (out stats, AF_INET6);
-			return new Win32IcmpV6Statistics (stats);
-		}
-
-		public override IPGlobalStatistics GetIPv4GlobalStatistics ()
-		{
-			if (!Socket.OSSupportsIPv4)
-				throw new NetworkInformationException ();
-			Win32_MIB_IPSTATS stats;
-			GetIpStatisticsEx (out stats, AF_INET);
-			return new Win32IPGlobalStatistics (stats);
-		}
-
-		public override IPGlobalStatistics GetIPv6GlobalStatistics ()
-		{
-			if (!Socket.OSSupportsIPv6)
-				throw new NetworkInformationException ();
-			Win32_MIB_IPSTATS stats;
-			GetIpStatisticsEx (out stats, AF_INET6);
-			return new Win32IPGlobalStatistics (stats);
-		}
-
-		public override TcpStatistics GetTcpIPv4Statistics ()
-		{
-			if (!Socket.OSSupportsIPv4)
-				throw new NetworkInformationException ();
-			Win32_MIB_TCPSTATS stats;
-			GetTcpStatisticsEx (out stats, AF_INET);
-			return new Win32TcpStatistics (stats);
-		}
-
-		public override TcpStatistics GetTcpIPv6Statistics ()
-		{
-			if (!Socket.OSSupportsIPv6)
-				throw new NetworkInformationException ();
-			Win32_MIB_TCPSTATS stats;
-			GetTcpStatisticsEx (out stats, AF_INET6);
-			return new Win32TcpStatistics (stats);
-		}
-
-		public override UdpStatistics GetUdpIPv4Statistics ()
-		{
-			if (!Socket.OSSupportsIPv4)
-				throw new NetworkInformationException ();
-			Win32_MIB_UDPSTATS stats;
-			GetUdpStatisticsEx (out stats, AF_INET);
-			return new Win32UdpStatistics (stats);
-		}
-
-		public override UdpStatistics GetUdpIPv6Statistics ()
-		{
-			if (!Socket.OSSupportsIPv6)
-				throw new NetworkInformationException ();
-			Win32_MIB_UDPSTATS stats;
-			GetUdpStatisticsEx (out stats, AF_INET6);
-			return new Win32UdpStatistics (stats);
-		}
-
-		public override string DhcpScopeName {
-			get { return Win32NetworkInterface.FixedInfo.ScopeId; }
-		}
-
-		public override string DomainName {
-			get { return Win32NetworkInterface.FixedInfo.DomainName; }
-		}
-
-		public override string HostName {
-			get { return Win32NetworkInterface.FixedInfo.HostName; }
-		}
-
-		public override bool IsWinsProxy {
-			get { return Win32NetworkInterface.FixedInfo.EnableProxy != 0; }
-		}
-
-		public override NetBiosNodeType NodeType {
-			get { return Win32NetworkInterface.FixedInfo.NodeType; }
-		}
-
-		// PInvokes
-
-		[DllImport ("iphlpapi.dll")]
-		static extern int GetTcpTable (byte [] pTcpTable, ref int pdwSize, bool bOrder);
-
-		[DllImport ("iphlpapi.dll")]
-		static extern int GetTcp6Table (byte [] TcpTable, ref int SizePointer, bool Order);
-
-		[DllImport ("iphlpapi.dll")]
-		static extern int GetUdpTable (byte [] pUdpTable, ref int pdwSize, bool bOrder);
-
-		[DllImport ("iphlpapi.dll")]
-		static extern int GetUdp6Table (byte [] Udp6Table, ref int SizePointer, bool Order);
-
-		[DllImport ("iphlpapi.dll")]
-		static extern int GetTcpStatisticsEx (out Win32_MIB_TCPSTATS pStats, int dwFamily);
-
-		[DllImport ("iphlpapi.dll")]
-		static extern int GetUdpStatisticsEx (out Win32_MIB_UDPSTATS pStats, int dwFamily);
-
-		[DllImport ("iphlpapi.dll")]
-		static extern int GetIcmpStatistics (out Win32_MIBICMPINFO pStats, int dwFamily);
-
-		[DllImport ("iphlpapi.dll")]
-		static extern int GetIcmpStatisticsEx (out Win32_MIB_ICMP_EX pStats, int dwFamily);
-
-		[DllImport ("iphlpapi.dll")]
-		static extern int GetIpStatisticsEx (out Win32_MIB_IPSTATS pStats, int dwFamily);
-
-		[DllImport ("Ws2_32.dll")]
-		static extern ushort ntohs (ushort netshort);
-
-		// Win32 structures
-
-		[StructLayout (LayoutKind.Explicit)]
-		struct Win32_IN6_ADDR
-		{
-			[FieldOffset (0)]
-			[MarshalAs ( UnmanagedType.ByValArray, SizeConst = 16)]
-			public byte [] Bytes;
-		}
-
-		[StructLayout (LayoutKind.Sequential)]
-		class Win32_MIB_TCPROW
-		{
-			public TcpState State;
-			public uint LocalAddr;
-			public int LocalPort;
-			public uint RemoteAddr;
-			public int RemotePort;
-
-			public IPEndPoint LocalEndPoint {
-				get { return new IPEndPoint (LocalAddr, LocalPort); }
-			}
-
-			public IPEndPoint RemoteEndPoint {
-				get { return new IPEndPoint (RemoteAddr, RemotePort); }
-			}
-
-			public TcpConnectionInformation TcpInfo {
-				get { return new SystemTcpConnectionInformation (LocalEndPoint, RemoteEndPoint, State); }
-			}
-		}
-
-		[StructLayout (LayoutKind.Sequential)]
-		class Win32_MIB_TCP6ROW
-		{
-			public TcpState State;
-			public Win32_IN6_ADDR LocalAddr;
-			public uint LocalScopeId;
-			public int LocalPort;
-			public Win32_IN6_ADDR RemoteAddr;
-			public uint RemoteScopeId;
-			public int RemotePort;
-
-			public IPEndPoint LocalEndPoint {
-				get { return new IPEndPoint (new IPAddress (LocalAddr.Bytes, LocalScopeId), LocalPort); }
-			}
-
-			public IPEndPoint RemoteEndPoint {
-				get { return new IPEndPoint (new IPAddress (RemoteAddr.Bytes, RemoteScopeId), RemotePort); }
-			}
-
-			public TcpConnectionInformation TcpInfo {
-				get { return new SystemTcpConnectionInformation (LocalEndPoint, RemoteEndPoint, State); }
-			}
-		}
-
-		[StructLayout (LayoutKind.Sequential)]
-		class Win32_MIB_UDPROW
-		{
-			public uint LocalAddr;
-			public uint LocalPort;
-
-			public IPEndPoint LocalEndPoint {
-				get { return new IPEndPoint (LocalAddr, ntohs((ushort)LocalPort)); }
-			}
-		}
-
-		[StructLayout (LayoutKind.Sequential)]
-		class Win32_MIB_UDP6ROW
-		{
-			public Win32_IN6_ADDR LocalAddr;
-			public uint LocalScopeId;
-			public int LocalPort;
-
-			public IPEndPoint LocalEndPoint {
-				get { return new IPEndPoint (new IPAddress (LocalAddr.Bytes, LocalScopeId), LocalPort); }
-			}
+#else
+			(new NetworkInformationPermission (NetworkInformationAccess.Read)).Demand ();
+			return new SystemIPGlobalProperties ();
+#endif
 		}
 	}
-#endif
 }
