+++ mcs/class/System/System.Net.NetworkInformation/UnixIPGlobalStatistics.cs
@@ -28,7 +28,6 @@
 //
 using System.Collections.Specialized;
 using System.Globalization;
-using System.Runtime.InteropServices;
 
 namespace System.Net.NetworkInformation {
 	class MibIPGlobalStatistics : IPGlobalStatistics
@@ -112,112 +111,4 @@ namespace System.Net.NetworkInformation {
 			get { return Get ("InUnknownProtos"); }
 		}
 	}
-
-#if WIN_PLATFORM
-	class Win32IPGlobalStatistics : IPGlobalStatistics 
-	{
-		Win32_MIB_IPSTATS info;
-
-		public Win32IPGlobalStatistics (Win32_MIB_IPSTATS info)
-		{
-			this.info = info;
-		}
-
-		public override int DefaultTtl {
-			get { return info.DefaultTTL; }
-		}
-		public override bool ForwardingEnabled {
-			get { return info.Forwarding != 0; }
-		}
-		public override int NumberOfInterfaces {
-			get { return info.NumIf; }
-		}
-		public override int NumberOfIPAddresses {
-			get { return info.NumAddr; }
-		}
-		public override int NumberOfRoutes {
-			get { return info.NumRoutes; }
-		}
-		public override long OutputPacketRequests {
-			get { return info.OutRequests; }
-		}
-		public override long OutputPacketRoutingDiscards {
-			get { return info.RoutingDiscards; }
-		}
-		public override long OutputPacketsDiscarded {
-			get { return info.OutDiscards; }
-		}
-		public override long OutputPacketsWithNoRoute {
-			get { return info.OutNoRoutes; }
-		}
-		public override long PacketFragmentFailures {
-			get { return info.FragFails; }
-		}
-		public override long PacketReassembliesRequired {
-			get { return info.ReasmReqds; }
-		}
-		public override long PacketReassemblyFailures {
-			get { return info.ReasmFails; }
-		}
-		public override long PacketReassemblyTimeout {
-			get { return info.ReasmTimeout; }
-		}
-		public override long PacketsFragmented {
-			get { return info.FragOks; }
-		}
-		public override long PacketsReassembled {
-			get { return info.ReasmOks; }
-		}
-		public override long ReceivedPackets {
-			get { return info.InReceives; }
-		}
-		public override long ReceivedPacketsDelivered {
-			get { return info.InDelivers; }
-		}
-		public override long ReceivedPacketsDiscarded {
-			get { return info.InDiscards; }
-		}
-		public override long ReceivedPacketsForwarded {
-			get { return info.ForwDatagrams; }
-		}
-		public override long ReceivedPacketsWithAddressErrors {
-			get { return info.InAddrErrors; }
-		}
-		public override long ReceivedPacketsWithHeadersErrors {
-			get { return info.InHdrErrors; }
-		}
-		public override long ReceivedPacketsWithUnknownProtocol {
-			get { return info.InUnknownProtos; }
-		}
-	}
-	
-	[StructLayout (LayoutKind.Sequential)]
-	struct Win32_MIB_IPSTATS
-	{
-		public int Forwarding;
-		public int DefaultTTL;
-		public uint InReceives;
-		public uint InHdrErrors;
-		public uint InAddrErrors;
-		public uint ForwDatagrams;
-		public uint InUnknownProtos;
-		public uint InDiscards;
-		public uint InDelivers;
-		public uint OutRequests;
-		public uint RoutingDiscards;
-		public uint OutDiscards;
-		public uint OutNoRoutes;
-		public uint ReasmTimeout;
-		public uint ReasmReqds;
-		public uint ReasmOks;
-		public uint ReasmFails;
-		public uint FragOks;
-		public uint FragFails;
-		public uint FragCreates;
-		public int NumIf;
-		public int NumAddr;
-		public int NumRoutes;
-	}
-#endif
 }
-
