+++ mcs/class/System/System.Net.NetworkInformation/UnixTcpStatistics.cs
@@ -87,92 +87,4 @@ namespace System.Net.NetworkInformation {
 			get { return Get ("OutSegs"); }
 		}
 	}
-
-#if WIN_PLATFORM
-	class Win32TcpStatistics : TcpStatistics
-	{
-		Win32_MIB_TCPSTATS info;
-
-		public Win32TcpStatistics (Win32_MIB_TCPSTATS info)
-		{
-			this.info = info;
-		}
-
-		public override long ConnectionsAccepted {
-			get { return info.PassiveOpens; }
-		}
-
-		public override long ConnectionsInitiated {
-			get { return info.ActiveOpens; }
-		}
-
-		public override long CumulativeConnections {
-			get { return info.NumConns; }
-		}
-
-		public override long CurrentConnections {
-			get { return info.CurrEstab; }
-		}
-
-		public override long ErrorsReceived {
-			get { return info.InErrs; }
-		}
-
-		public override long FailedConnectionAttempts {
-			get { return info.AttemptFails; }
-		}
-
-		public override long MaximumConnections {
-			get { return info.MaxConn; }
-		}
-
-		public override long MaximumTransmissionTimeout {
-			get { return info.RtoMax; }
-		}
-
-		public override long MinimumTransmissionTimeout {
-			get { return info.RtoMin; }
-		}
-
-		public override long ResetConnections {
-			get { return info.EstabResets; }
-		}
-
-		public override long ResetsSent {
-			get { return info.OutRsts; }
-		}
-
-		public override long SegmentsReceived {
-			get { return info.InSegs; }
-		}
-
-		public override long SegmentsResent {
-			get { return info.RetransSegs; }
-		}
-
-		public override long SegmentsSent {
-			get { return info.OutSegs; }
-		}
-	}
-
-	struct Win32_MIB_TCPSTATS
-	{
-		public uint RtoAlgorithm;
-		public uint RtoMin;
-		public uint RtoMax;
-		public uint MaxConn;
-		public uint ActiveOpens;
-		public uint PassiveOpens;
-		public uint AttemptFails;
-		public uint EstabResets;
-		public uint CurrEstab;
-		public uint InSegs;
-		public uint OutSegs;
-		public uint RetransSegs;
-		public uint InErrs;
-		public uint OutRsts;
-		public uint NumConns;
-	}
-#endif
 }
-
