+++ mcs/class/System/System.Net.NetworkInformation/Win32IPAddressCollection.cs
@@ -26,15 +26,10 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Net;
+#if WIN_PLATFORM
 using System.Runtime.InteropServices;
 
 namespace System.Net.NetworkInformation {
-
-#if WIN_PLATFORM
 	class Win32IPAddressCollection : IPAddressCollection
 	{
 		public static readonly Win32IPAddressCollection Empty = new Win32IPAddressCollection (IntPtr.Zero);
@@ -93,7 +88,5 @@ namespace System.Net.NetworkInformation {
 			}
 		}
 	}
-#endif
 }
-
-
+#endif
