+++ mcs/class/System/System.Net.NetworkInformation/Win32IPv6InterfaceProperties.cs
@@ -1,0 +1,49 @@
+//
+// System.Net.NetworkInformation.IPv6InterfaceProperties
+//
+// Authors:
+//	Gonzalo Paniagua Javier (gonzalo@novell.com)
+//	Atsushi Enomoto (atsushi@ximian.com)
+//
+// Copyright (c) 2006-2007 Novell, Inc. (http://www.novell.com)
+//
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject to
+// the following conditions:
+// 
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+#if WIN_PLATFORM
+namespace System.Net.NetworkInformation {
+	class Win32IPv6InterfaceProperties : IPv6InterfaceProperties
+	{
+		Win32_MIB_IFROW mib;
+
+		public Win32IPv6InterfaceProperties (Win32_MIB_IFROW mib)
+		{
+			this.mib = mib;
+		}
+
+		public override int Index {
+			get { return mib.Index; }
+		}
+
+		public override int Mtu {
+			get { return mib.Mtu; }
+		}
+	}
+}
+#endif
