+++ mcs/class/System/System.Net.NetworkInformation/Win32IcmpV6Statistics.cs
@@ -26,125 +26,11 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-using System.Collections.Specialized;
-using System.Globalization;
+#if WIN_PLATFORM
 using System.Runtime.InteropServices;
 
 namespace System.Net.NetworkInformation {
-	class MibIcmpV6Statistics : IcmpV6Statistics
-	{
-		StringDictionary dic;
-
-		public MibIcmpV6Statistics (StringDictionary dic)
-		{
-			this.dic = dic;
-		}
-
-		long Get (string name)
-		{
-			return dic [name] != null ? long.Parse (dic [name], NumberFormatInfo.InvariantInfo) : 0;
-		}
-
-		public override long DestinationUnreachableMessagesReceived {
-			get { return Get ("InDestUnreachs"); }
-		}
-		public override long DestinationUnreachableMessagesSent {
-			get { return Get ("OutDestUnreachs"); }
-		}
-		public override long EchoRepliesReceived {
-			get { return Get ("InEchoReplies"); }
-		}
-		public override long EchoRepliesSent {
-			get { return Get ("OutEchoReplies"); }
-		}
-		public override long EchoRequestsReceived {
-			get { return Get ("InEchos"); }
-		}
-		public override long EchoRequestsSent {
-			get { return Get ("OutEchos"); }
-		}
-		public override long ErrorsReceived {
-			get { return Get ("InErrors"); }
-		}
-		public override long ErrorsSent {
-			get { return Get ("OutErrors"); }
-		}
-		public override long MembershipQueriesReceived {
-			get { return Get ("InGroupMembQueries"); }
-		}
-		public override long MembershipQueriesSent {
-			get { return Get ("OutGroupMembQueries"); }
-		}
-		public override long MembershipReductionsReceived {
-			get { return Get ("InGroupMembReductiions"); }
-		}
-		public override long MembershipReductionsSent {
-			get { return Get ("OutGroupMembReductiions"); }
-		}
-		public override long MembershipReportsReceived {
-			get { return Get ("InGroupMembRespons"); }
-		}
-		public override long MembershipReportsSent {
-			get { return Get ("OutGroupMembRespons"); }
-		}
-		public override long MessagesReceived {
-			get { return Get ("InMsgs"); }
-		}
-		public override long MessagesSent {
-			get { return Get ("OutMsgs"); }
-		}
-		public override long NeighborAdvertisementsReceived {
-			get { return Get ("InNeighborAdvertisements"); }
-		}
-		public override long NeighborAdvertisementsSent {
-			get { return Get ("OutNeighborAdvertisements"); }
-		}
-		public override long NeighborSolicitsReceived {
-			get { return Get ("InNeighborSolicits"); }
-		}
-		public override long NeighborSolicitsSent {
-			get { return Get ("OutNeighborSolicits"); }
-		}
-		public override long PacketTooBigMessagesReceived {
-			get { return Get ("InPktTooBigs"); }
-		}
-		public override long PacketTooBigMessagesSent {
-			get { return Get ("OutPktTooBigs"); }
-		}
-		public override long ParameterProblemsReceived {
-			get { return Get ("InParmProblems"); }
-		}
-		public override long ParameterProblemsSent {
-			get { return Get ("OutParmProblems"); }
-		}
-		public override long RedirectsReceived {
-			get { return Get ("InRedirects"); }
-		}
-		public override long RedirectsSent {
-			get { return Get ("OutRedirects"); }
-		}
-		public override long RouterAdvertisementsReceived {
-			get { return Get ("InRouterAdvertisements"); }
-		}
-		public override long RouterAdvertisementsSent {
-			get { return Get ("OutRouterAdvertisements"); }
-		}
-		public override long RouterSolicitsReceived {
-			get { return Get ("InRouterSolicits"); }
-		}
-		public override long RouterSolicitsSent {
-			get { return Get ("OutRouterSolicits"); }
-		}
-		public override long TimeExceededMessagesReceived {
-			get { return Get ("InTimeExcds"); }
-		}
-		public override long TimeExceededMessagesSent {
-			get { return Get ("OutTimeExcds"); }
-		}
-	}
-
-	class IcmpV6MessageTypes
-	{
+	class IcmpV6MessageTypes {
 		public const int DestinationUnreachable = 1;
 		public const int PacketTooBig = 2;
 		public const int TimeExceeded = 3;
@@ -162,7 +48,6 @@ namespace System.Net.NetworkInformation {
 		public const int RouterRenumbering = 138;
 	}
 
-#if WIN_PLATFORM
 	class Win32IcmpV6Statistics : IcmpV6Statistics
 	{
 		Win32_MIBICMPSTATS_EX iin, iout;
@@ -284,6 +169,5 @@ namespace System.Net.NetworkInformation {
 		[MarshalAs (UnmanagedType.ByValArray, SizeConst = 256)]
 		public uint [] Counts;
 	}
-#endif
 }
-
+#endif
