+++ mcs/class/System/System.Net.NetworkInformation/Win32NetworkInterfaceFactory.cs
@@ -30,63 +30,17 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-using System;
-using System.Collections.Generic;
-using System.Collections;
-using System.Net;
-using System.Net.Sockets;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.IO;
-using System.Globalization;
-
+#if WIN_PLATFORM
 namespace System.Net.NetworkInformation {
-	static class SystemNetworkInterface {
-
-		static readonly NetworkInterfaceFactory nif = NetworkInterfaceFactory.Create ();
-
-		public static NetworkInterface [] GetNetworkInterfaces ()
-		{
-			try {
-				return nif.GetAllNetworkInterfaces ();
-			} catch {
-				return new NetworkInterface [0];
-			}
-		}
-
-		public static bool InternalGetIsNetworkAvailable ()
-		{
-			// TODO:
-			return true;
-		}
-
-		public static int InternalLoopbackInterfaceIndex {
-			get {
-				return nif.GetLoopbackInterfaceIndex ();
-			}
-		}
-
-		public static int InternalIPv6LoopbackInterfaceIndex {
-			get {
-				throw new NotImplementedException ();
-			}
-		}
-
-		public static IPAddress GetNetMask (IPAddress address)
-		{
-			return nif.GetNetMask (address);
-		}
-	}
-
-	abstract class NetworkInterfaceFactory
-	{
-		public abstract NetworkInterface [] GetAllNetworkInterfaces ();
-		public abstract int GetLoopbackInterfaceIndex ();
-		public abstract IPAddress GetNetMask (IPAddress address);
-
+	internal static class Win32NetworkInterfaceFactoryPal {
 		public static NetworkInterfaceFactory Create ()
 		{
-			return NetworkInterfaceFactoryPal.Create ();
+			Version windowsVer51 = new Version (5, 1);
+			if (Environment.OSVersion.Version >= windowsVer51)
+				return new Win32NetworkInterfaceAPI ();
+
+			return null;
 		}
 	}
 }
+#endif
