+++ mcs/class/System/System.csproj
@@ -638,6 +638,7 @@
     <Compile Include="..\referencesource\System\net\System\Net\NetworkInformation\NetworkInformationException.cs" />
     <Compile Include="..\referencesource\System\net\System\Net\NetworkInformation\NetworkInformationPermission.cs" />
     <Compile Include="..\referencesource\System\net\System\Net\NetworkInformation\NetworkInterface.cs" />
+    <Compile Include="..\referencesource\System\net\System\Net\NetworkInformation\NetworkInterfaceFactory.cs" />
     <Compile Include="..\referencesource\System\net\System\Net\NetworkInformation\NetworkInterfaceComponent.cs" />
     <Compile Include="..\referencesource\System\net\System\Net\NetworkInformation\OperationalStatus.cs" />
     <Compile Include="..\referencesource\System\net\System\Net\NetworkInformation\PhysicalAddress.cs" />
@@ -877,26 +878,48 @@
     <Compile Include="System.Net.Mail\SmtpFailedRecipientException.cs" />
     <Compile Include="System.Net.Mail\SmtpFailedRecipientsException.cs" />
     <Compile Include="System.Net.Mail\SmtpStatusCode.cs" />
-    <Compile Include="System.Net.NetworkInformation\IPAddressCollection.cs" />
-    <Compile Include="System.Net.NetworkInformation\IPGlobalProperties.cs" />
-    <Compile Include="System.Net.NetworkInformation\IPGlobalStatistics.cs" />
-    <Compile Include="System.Net.NetworkInformation\IPInterfaceProperties.cs" />
-    <Compile Include="System.Net.NetworkInformation\IPv4InterfaceProperties.cs" />
-    <Compile Include="System.Net.NetworkInformation\IPv4InterfaceStatistics.cs" />
-    <Compile Include="System.Net.NetworkInformation\IPv6InterfaceProperties.cs" />
-    <Compile Include="System.Net.NetworkInformation\IcmpV4Statistics.cs" />
-    <Compile Include="System.Net.NetworkInformation\IcmpV6Statistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\LinuxIPInterfaceProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\LinuxIPv4InterfaceProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\LinuxIPv4InterfaceStatistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\LinuxNetworkInterface.cs" />
     <Compile Include="System.Net.NetworkInformation\LinuxNetworkInterfaceMarshal.cs" />
+    <Compile Include="System.Net.NetworkInformation\LinuxUnicastIPAddressInformation.cs" />
+    <Compile Include="System.Net.NetworkInformation\MacOsIPInterfaceProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\MacOsIPv4InterfaceProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\MacOsIPv4InterfaceStatistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\MacOsNetworkInterface.cs" />
     <Compile Include="System.Net.NetworkInformation\MacOsNetworkInterfaceMarshal.cs" />
+    <Compile Include="System.Net.NetworkInformation\IPGlobalPropertiesFactory.cs" />
     <Compile Include="System.Net.NetworkInformation\NetworkChange.cs" />
     <Compile Include="System.Net.NetworkInformation\NetworkInterface.cs" />
     <Compile Include="System.Net.NetworkInformation\Ping.cs" />
     <Compile Include="System.Net.NetworkInformation\PingCompletedEventArgs.cs" />
     <Compile Include="System.Net.NetworkInformation\PingCompletedEventHandler.cs" />
-    <Compile Include="System.Net.NetworkInformation\TcpStatistics.cs" />
-    <Compile Include="System.Net.NetworkInformation\UdpStatistics.cs" />
-    <Compile Include="System.Net.NetworkInformation\UnicastIPAddressInformation.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixIcmpV4Statistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixIcmpV6Statistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixIPGlobalProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixIPGlobalStatistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixIPInterfaceProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixIPv4InterfaceProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixNetworkInterface.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixNetworkInterfaceFactory.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixTcpStatistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\UnixUdpStatistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32IcmpV4Statistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32IcmpV6Statistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32IPAddressCollection.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32IPGlobalProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32IPGlobalStatistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32IPInterfaceProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32IPv4InterfaceProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32IPv4InterfaceStatistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32IPv6InterfaceProperties.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32NetworkInterface.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32NetworkInterfaceFactory.cs" />
     <Compile Include="System.Net.NetworkInformation\Win32NetworkInterfaceMarshal.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32TcpStatistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32UdpStatistics.cs" />
+    <Compile Include="System.Net.NetworkInformation\Win32UnicastIPAddressInformation.cs" />
     <Compile Include="System.Net.Security\EncryptionPolicy.cs" />
     <Compile Include="System.Net.Security\LocalCertificateSelectionCallback.cs" />
     <Compile Include="System.Net.Security\NegotiateStream.cs" />
