+++ mcs/class/System/linux_networkinfo.sources
@@ -1,0 +1,6 @@
+System.Net.NetworkInformation/LinuxIPInterfaceProperties.cs
+System.Net.NetworkInformation/LinuxIPv4InterfaceProperties.cs
+System.Net.NetworkInformation/LinuxIPv4InterfaceStatistics.cs
+System.Net.NetworkInformation/LinuxNetworkInterfaceMarshal.cs
+System.Net.NetworkInformation/LinuxNetworkInterface.cs
+System.Net.NetworkInformation/LinuxUnicastIPAddressInformation.cs
