+++ mcs/class/System/macos_networkinfo.sources
@@ -1,0 +1,5 @@
+System.Net.NetworkInformation/MacOsIPInterfaceProperties.cs
+System.Net.NetworkInformation/MacOsIPv4InterfaceProperties.cs
+System.Net.NetworkInformation/MacOsIPv4InterfaceStatistics.cs
+System.Net.NetworkInformation/MacOsNetworkInterfaceMarshal.cs
+System.Net.NetworkInformation/MacOsNetworkInterface.cs
