+++ mcs/class/System/unix_networkinfo.sources
@@ -1,0 +1,10 @@
+System.Net.NetworkInformation/UnixIcmpV4Statistics.cs
+System.Net.NetworkInformation/UnixIcmpV6Statistics.cs
+System.Net.NetworkInformation/UnixIPGlobalProperties.cs
+System.Net.NetworkInformation/UnixIPGlobalStatistics.cs
+System.Net.NetworkInformation/UnixIPInterfaceProperties.cs
+System.Net.NetworkInformation/UnixIPv4InterfaceProperties.cs
+System.Net.NetworkInformation/UnixNetworkInterface.cs
+System.Net.NetworkInformation/UnixNetworkInterfaceFactory.cs
+System.Net.NetworkInformation/UnixTcpStatistics.cs
+System.Net.NetworkInformation/UnixUdpStatistics.cs
