+++ mcs/class/System/winaot_System.dll.exclude.sources
@@ -1,0 +1,3 @@
+#include unix_networkinfo.sources
+#include linux_networkinfo.sources
+#include macos_networkinfo.sources
