+++ mcs/class/System/winaot_System.dll.sources
@@ -1,1 +1,3 @@
 #include mobile_System.dll.sources
+
+System.Net.NetworkInformation/Win32UnixFactoryPal.cs
